refactor(dh): move top bar store to util-top-bar library
