+++ libs/dh/core/shell/src/lib/dh-core-shell.component.ts
@@ -23,9 +23,9 @@ import { PushModule } from '@rx-angular/template';
 import { WattButtonModule, WattShellComponent } from '@energinet-datahub/watt';
 import { DhLanguagePickerModule } from '@energinet-datahub/dh/globalization/feature-language-picker';
 import { DhApplicationInsights } from '@energinet-datahub/dh/shared/util-application-insights';
+import { DhTopBarStore } from '@energinet-datahub/dh/shared/util-top-bar';
 
 import { DhPrimaryNavigationScam } from './dh-primary-navigation.component';
-import { DhTopBarStore } from './dh-top-bar.store';
 
 @Component({
   selector: 'dh-shell',
