+++ libs/dh/shared/util-top-bar/src/lib/default-title-translation-key.ts
@@ -1,0 +1,1 @@
+export const defaultTitleTranslationKey = '';
