--- libs/dh/shared/util-top-bar/src/lib/dh-shared-util-top-bar.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  imports: [CommonModule],
-})
-export class DhSharedUtilTopBarModule {}
